Remove duplicate combat definitions from main orchestrator
