+++ src/data/index.js
@@ -2,3 +2,6 @@ export * from "./backgrounds.js";
 export * from "./sprites.js";
 export * from "./bestiary.js";
 export * from "./items.js";
+export * from "./memories.js";
+export * from "./enemies.js";
+export * from "./rooms.js";
