+++ src/state/config.js
@@ -1,0 +1,10 @@
+export const MAX_CONSUMABLE_SLOTS = 3;
+export const MERCHANT_BASE_DRAFT_COST = 10;
+export const MERCHANT_DRAFT_COST_INCREMENT = 5;
+
+export const DEFAULT_PLAYER_STATS = {
+  maxEssence: 15,
+  baseApRegen: 6,
+  baseCritChance: 5,
+  apCarryover: 12,
+};
